+++ CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+0.4.4
+*****
+* Added the ``EnenergyGatherer()`` class, a Mapping for managing all
+  forcefield potential energies.
+* Changed the minimum Pandas version to ``>= 0.24.0``.
+* Updated the keyword arguments of ``get_asa_md()``.
+* Fixed an issue were the charge was not properly set to an integer value
+  when optimizing individual ligands.
+
+
 0.4.3
 *****
 * Changed the a number of function signatures in md_asa.py to ensure signature